lombok delete, add services

+++ src/main/resources/db/migration/V2__Create_locations_table.sql
@@ -1,8 +1,8 @@
 CREATE TABLE locations
 (
-    id        SERIAL PRIMARY KEY,
+    id        BIGSERIAL PRIMARY KEY,
     name      VARCHAR(255)  NOT NULL,
-    user_id   INT REFERENCES users (id) ON DELETE CASCADE,
+    user_id   BIGSERIAL REFERENCES users (id) ON DELETE CASCADE,
     latitude  DECIMAL(9, 6) NOT NULL,
     longitude DECIMAL(9, 6) NOT NULL
 );
