+++ src/main/resources/db/migration/V3__Create_sessions_table.sql
@@ -1,6 +1,6 @@
 CREATE TABLE sessions
 (
     id         UUID PRIMARY KEY DEFAULT gen_random_uuid(),
-    user_id    INT REFERENCES users (id) ON DELETE CASCADE,
+    user_id    BIGSERIAL REFERENCES users (id) ON DELETE CASCADE,
     expires_at TIMESTAMP NOT NULL
 );
